Implemented Vlads suggested edits
