+++ .gitignore
@@ -10,3 +10,4 @@ apps/*/.idea
 */**/yarn-debug.log*
 */**/yarn-error.log*
 
+.DS_Store
