+++ apps/jira-node/.env
@@ -1,6 +1,6 @@
 PORT=8302
 MONDAY_SIGNING_SECRET=
-TUNNEL_SUBDOMAIN=jql-app-803
+TUNNEL_SUBDOMAIN=
 TUNNEL_SERVER_HOST=http://loca.lt
