+++ apps/jira-node/.eslintrc.json
