+++ apps/jira-node/.gitignore
@@ -1,5 +1,8 @@
 node_modules/
 *.log
+.editorconfig
+database.sqlite3
+.DS_Store
 
 # Created by https://www.gitignore.io/api/node,rubymine,visualstudiocode
 # Edit at https://www.gitignore.io/?templates=node,rubymine,visualstudiocode
