+++ apps/jira-node/.sequelizerc
