+++ apps/jira-node/LICENSE
