+++ apps/jira-node/src/db/config/db.config.json
