+++ apps/jira-node/src/db/models/subscription.js
@@ -2,11 +2,7 @@
 const { Model } = require('sequelize');
 
 module.exports = (sequelize, DataTypes) => {
-  class Subscription extends Model {
-    static associate(models) {
-      // define association here
-    }
-  }
+  class Subscription extends Model{}
   Subscription.init(
     {
       active: DataTypes.BOOLEAN,
