+++ apps/jira-node/src/middlewares/authentication.js
